Pass props through routes, only one base table to customize
