+++ package.json
@@ -5,7 +5,6 @@
   "main": "app.js",
   "scripts": {
     "start": "node app.js",
-    "start:dev": "webpack-dev-server --inline --content-base . --history-api-fallback",
     "test": "node ./node_modules/mocha/bin/mocha --recursive --require babel-core/register test",
     "test_cover": "node ./node_modules/babel-istanbul/lib/cli.js --hook-run-in-context cover node_modules/mocha/bin/_mocha -- -R spec --timeout 10000 --recursive test",
     "webpack": "node ./node_modules/webpack/bin/webpack.js"
