+++ src/client/main.jsx
@@ -3,11 +3,11 @@ import React from 'react';
 import {render} from 'react-dom';
 import App from './modules/app.jsx';
 import {Router, Route, hashHistory, IndexRedirect} from 'react-router';
-import ProjectsPage from './modules/content/pages/ProjectsPage.jsx';
 import UserProfilePage from './modules/content/pages/UserProfilePage.jsx';
 import OrganizationsPage from './modules/content/pages/OrganizationsPage.jsx';
 import ContentWrapper from './modules/content/ContentWrapper.jsx';
 import ProjectPage from './modules/content/pages/ProjectPage.jsx';
+import ProjectsPage from './modules/content/pages/ProjectsPage.jsx';
 
 require('bootstrap-webpack');
 require('font-awesome-webpack');
