+++ src/client/modules/content/ContentWrapper.jsx
@@ -1,16 +1,19 @@
 /* global window */
 import React from 'react';
-import {Router, Route, hashHistory, browserHistory} from 'react-router';
-import ProjectsPage from './pages/ProjectsPage.jsx';
 
 export default class ContentWrapper extends React.Component {
 
     constructor(props) {
         super(props);
+        this.restClient = this.props.restClient;
     }
 
     render() {
-        
+
+        let PageWithRestClient = React.cloneElement(this.props.children, {
+            restClient: this.restClient
+        });
+
 
         return <div className="content-wrapper">
 
@@ -28,7 +31,7 @@ export default class ContentWrapper extends React.Component {
                 </ol>
             </section>
 
-            {this.props.children || <ProjectsPage/>}
+            {PageWithRestClient}
 
         </div>;
     }
